+++ src/main/java/org/broadinstitute/consent/http/models/DatasetEntry.java
@@ -1,0 +1,36 @@
+package org.broadinstitute.consent.http.models;
+
+public class DatasetEntry {
+
+    String key;
+    String value;
+    String label;
+
+    public DatasetEntry() {
+    }
+
+    public String getKey() {
+        return key;
+    }
+
+    public void setKey(String key) {
+        this.key = key;
+    }
+
+    public String getValue() {
+        return value;
+    }
+
+    public void setValue(String value) {
+        this.value = value;
+    }
+
+    public String getLabel() {
+        return label;
+    }
+
+    public void setLabel(String label) {
+        this.label = label;
+    }
+
+}
